Fix for overrides wrapping from minimum to maximum rate
Overrides can wrap around to the maximum rate if multiple CMD_OVERRIDE_xx_MINUS commands are queued while
already at, or near to, the minimum rate. This can occur when using an encoder if moved too quickly.

This happens because the arguments to plan_feed_override() and spindle_set_override() are unsigned. If
(for instance) the override rate is 10, and commands from 11 negative encoder ticks are enqueued before
processing, then the requested override wraps round to a large positive number, and is capped at the
max override value.