Shift neopixel bit pattern
On some processors, MOSI can be driven to the output level before
the clock is enabled. If the first bit is a 1, this can result in
a longer pulse that is interpreted as a high level by neopixels.

This fix maintains the same inter-bit timings, but ensures the
first bit is always 0. (The last bit is already always zero due
to padding for reset timing requirements).